ACOMMONS-14: Move SonarLintCache into org.sonar.plugins.api
