+++ commons/src/test/java/org/sonar/plugins/commons/api/SonarLintCacheTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.commons;
+package org.sonar.plugins.commons.api;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
